+++ src/csscn/__init__.py
@@ -1,0 +1,6 @@
+"""服务器安全卫士(原生版)模块"""
+
+from .client import CSSCNClient
+from .commands import csscn
+
+__all__ = ['CSSCNClient', 'csscn']